+++ testsuite/tests/overloadedrecflds/should_compile/T22424.hs
@@ -22,7 +22,8 @@ $(do
     x1 <- newName "x1"
     x2 <- newName "x2"
     let expr = UInfixE (VarE fld1 `AppE` VarE x1) (VarE '(&&)) (VarE fld2 `AppE` VarE x2)
-        fun_decl = FunD fun [Clause [VarP x1, VarP x2] (NormalB expr) []]
+        pats = [VarP x1, VarP x2]
+        fun_decl = FunD fun [Clause (map VisAP pats) (NormalB expr) []]
     pure [r1,r2,fun_decl]
  )
 
@@ -35,6 +36,7 @@ $(do
     x1 <- newName "x1"
     x2 <- newName "x2"
     let expr = UInfixE (VarE fld1 `AppE` VarE x1) (VarE '(&&)) (VarE fld2 `AppE` VarE x2)
-        fun_decl = FunD fun [Clause [VarP x1, VarP x2] (NormalB expr) []]
+        pats = [VarP x1, VarP x2]
+        fun_decl = FunD fun [Clause (map VisAP pats) (NormalB expr) []]
     pure [r1,r2,fun_decl]
  )