+++ testsuite/tests/th/T3899a.hs
@@ -10,6 +10,6 @@ data Nil = Nil
 
 nestedTuple n = do
      xs <- replicateM n (newName "x")
-     return $ LamE [foldr (\v prev -> ParensP (ConP 'Cons [] [VarP v,prev]))
+     return $ LamE [VisAP $ foldr (\v prev -> ParensP (ConP 'Cons [] [VarP v,prev]))
                        (ConP 'Nil [] []) xs]
                    (TupE $ map (Just . VarE) xs)